analysisServer/requirements.txt
opencv_python==4.5.5.64
Keras==2.9.0
pandas==0.25.3
numpy==1.20.0
imutils==0.5.3
scikit_learn==0.22.1
tensorflow==2.9.1
flask_caching==2.0.1